refactor(build-optimizer): add BuildOptimizerContainer and BuildOptimizerGrid components
